metric_learning_authentication_code/deep_puf_release/requirements.txt
lightning>2.0.0
omegaconf==2.3.0
wandb
jaxtyping
typeguard
Pillow
matplotlib
opencv-python
imageio
torchdyn
einops
tensorboard
tensorboardX
scikit-learn
tqdm
seaborn
scipy
openpyxl